4주차 과제

+++ main/views.py
@@ -7,14 +7,18 @@ def mainpage(request):
 def secondpage(request):
     return render(request, 'main/secondpage.html')
 def new_post(request):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
     return render(request, 'main/new_post.html')
 
 
 def create(request):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
     new_post = Post()
 
     new_post.title = request.POST['title']
-    new_post.writer = request.POST['writer']
+    new_post.writer = request.user.username
     new_post.content = request.POST['content']
     new_post.pub_date = request.POST['pub_date']
     new_post.category = request.POST['category']
@@ -32,13 +36,26 @@ def detail(request, post_id):
     return render(request, 'main/detail.html', {'post': post})
 
 def edit(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
     edit_post = get_object_or_404(Post, id=post_id)
+
+    if edit_post.writer != request.user.username:
+        return redirect('main:detail', edit_post.id)
+
     return render(request, 'main/edit.html', {'post': edit_post})
 
 def update(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
     update_post = get_object_or_404(Post, id=post_id)
+    if update_post.writer != request.user.username:
+        return redirect('main:detail', update_post.id)
+
     update_post.title = request.POST['title']
-    update_post.writer = request.POST['writer']
+    update_post.writer = request.user.username
     update_post.content = request.POST['content']
     update_post.pub_date = request.POST['pub_date']
     update_post.category = request.POST['category']
@@ -46,7 +63,14 @@ def update(request, post_id):
     return redirect('main:detail', update_post.id)
 
 def delete(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('accounts:login')
+
     delete_post = get_object_or_404(Post, id=post_id)
+
+    if delete_post.writer != request.user.username:
+        return redirect('main:detail', delete_post.id)
+
     delete_post.delete()
 
     return redirect('main:postpage')
